Reverted some changes as per review comments
